- split out shared helpers + types
